WIP: Start refactoring into services
