+++ packages/framework-provider-local-infrastructure/src/controllers/auth.ts
@@ -1,5 +1,5 @@
 import * as express from 'express'
-import { RuntimeStorage } from '../runtime-storage'
+import { UserRegistry } from '../services/user-registry'
 import { UserApp, UserEnvelope, BoosterConfig, NotAuthorizedError } from '@boostercloud/framework-types'
 import { UUID } from '@boostercloud/framework-types'
 
@@ -11,15 +11,15 @@ import { UUID } from '@boostercloud/framework-types'
 export class AuthController {
   public router: express.Router = express.Router()
 
-  constructor(readonly storage: RuntimeStorage, readonly config: BoosterConfig, readonly userProject: UserApp) {
+  constructor(readonly userRegistry: UserRegistry, readonly config: BoosterConfig, readonly userProject: UserApp) {
     this.router.post('/sign-up', async (req: express.Request, res: express.Response) => {
-      await this.signUp(req.body)
+      await this.userRegistry.signUp(req.body)
       res.status(200)
     })
 
     this.router.post('/sign-in', async (req: express.Request, res: express.Response) => {
       try {
-        const token = await this.signIn(req.body)
+        const token = await this.userRegistry.signIn(req.body)
         res.status(200).json(token)
       } catch (e) {
         if (e.name == NotAuthorizedError.name) {
@@ -30,33 +30,13 @@ export class AuthController {
     })
 
     this.router.post('/sign-out', async (req: express.Request, res: express.Response) => {
-      await this.signOut(req.body)
-      res.status(200)
+      const token = req.body?.accessToken
+      if (token) {
+        await this.userRegistry.signOut(token)
+        res.status(200)
+      } else {
+        res.status(400).json('accessToken field not set')
+      }
     })
   }
-
-  public async signUp(user: UserEnvelope): Promise<void> {
-    await this.userProject.boosterPreSignUpChecker(user)
-    await this.storage.registerUser(user)
-  }
-
-  public async signIn(user: UserEnvelope): Promise<UUID> {
-    await this.userProject.boosterPreSignUpChecker(user)
-    const token = UUID.generate()
-    const registeredMatches = await this.storage.getRegisteredUsersByEmail(user.email)
-    if (registeredMatches.length === 0)
-      throw new NotAuthorizedError(`User with email ${user.email} has not been registered `)
-    await this.storage.authenticateUser(token, user)
-    return token
-  }
-
-  public async signOut(request: SignOutBody): Promise<void> {
-    const token = request.accessToken
-    if (!token) throw new NotAuthorizedError('No access token provided in the body of the request')
-    await this.storage.signOutUser(token)
-  }
-}
-
-interface SignOutBody {
-  accessToken?: UUID
 }
